Upgraded groovy and groovy compilers

+++ src/test/groovy/com/cedarsoftware/util/io/TestAtomicInteger.groovy
@@ -91,7 +91,7 @@ class TestAtomicInteger
         TestUtil.printLine("json=" + json)
         list = (List) TestUtil.readJsonObject(json)
         assert list.size() == 2
-        atomicInt = list.get(0)
+        atomicInt = (AtomicInteger)list.get(0)
         assert atomicInt.get() == new AtomicInteger(12345).get()
         assertNotSame(list.get(0), list.get(1))
     }
